답안1(try4) solve
계산할때 int 범위 초과하는 부분 변경
j-i의 최대값은 50000인데 50000*50001은 int 범위를 초과함

+++ src/com/algorithm/training/two_pointers/subarray_product_less_than_k/Solution.java
@@ -5,7 +5,7 @@ public class Solution {
         int subArrayCount = 0, i = 0, j = 0, multi = 1;
         while(j <= nums.length) {
             if(j == nums.length) {
-                subArrayCount +=  ((j-i)*(j-i+1) / 2);
+                subArrayCount +=  ((long)(j-i)*(j-i+1) / 2);
             } else {
                 multi *= nums[j];
                 while(multi >= k && i <= j) {
